+++ src/main/java/com/accenture/concrete/ConcreteApplication.java
@@ -2,8 +2,18 @@ package com.accenture.concrete;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.cache.annotation.EnableCaching;
 
+import com.accenture.concrete.properties.CorsProperties;
+import com.accenture.concrete.properties.CouponProperties;
+
+@EnableCaching
 @SpringBootApplication
+@EnableConfigurationProperties({
+	CouponProperties.class,
+	CorsProperties.class
+})
 public class ConcreteApplication {
 
 	public static void main(String[] args) {