+++ src/main/java/com/accenture/concrete/service/CouponAPI.java
@@ -1,0 +1,11 @@
+package com.accenture.concrete.service;
+
+import java.util.List;
+
+import com.accenture.concrete.domain.Coupon;
+
+public interface CouponAPI {
+
+	public List<Coupon> findAll();
+	
+}