+++ src/test/java/com/accenture/concrete/ConcreteApplicationTests.java
@@ -1,5 +1,6 @@
 package com.accenture.concrete;
 
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -7,6 +8,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 class ConcreteApplicationTests {
 
 	@Test
+	@DisplayName("context loaded")
 	void contextLoads() {
 	}